requirements.txt
apache-airflow-providers-postgres==5.6.0
psycopg2-binary==2.9.6
python-dotenv==1.0.1
pandas==1.5.3 
scikit-learn==1.3.0
requests==2.32.3